consistency-checker/requirements.txt
flask
flask-cors
pdfkit
nltk
requests
python-dotenv
gunicorn
